fix clean.sh

+++ script/clean.sh
@@ -1,1 +1,5 @@
-rm /tmp/HelloWorld-assembly-1.0.jar
+#!/bin/sh
+
+if [ -e /tmp/HelloWorld-assembly-1.0.jar ]; then
+  rm /tmp/HelloWorld-assembly-1.0.jar
+fi
